bucket stuff and removed templates for math because
it made everything worse, the overloaded operators piced up weird emum
additions in windows.h

+++ kd_tree.hpp
@@ -1,25 +1,14 @@
 #pragma once
 
-#include <type_traits>
 #include "arraylist.hpp"
 #include "math.hpp"
 #include "soa_sort.hpp"
-#include <vcruntime_typeinfo.h>
 
-
-/**
- * An axis aligned (bounding) box data structures used for search queries.
- */
-struct AxisAlignedBox {
+struct Axis_Aligned_Box {
     /// The minimum and maximum coordinate corners of the 3D cuboid.
     V3 min {};
     V3 max {};
 
-    /**
-     * Tests whether the axis aligned box contains a point.
-     * @param pt The point.
-     * @return True if the box contains the point.
-     */
     bool contains(const V3 pt) const {
         if (pt.x >= min.x && pt.y >= min.y && pt.z >= min.z &&
             pt.x <= max.x && pt.y <= max.y && pt.z <= max.z)
@@ -30,23 +19,10 @@ struct AxisAlignedBox {
 
 struct Empty {};
 
-
-
-struct Pure_Kd_Tree {
-
-};
-
-/**
- * The k-d-tree class. Used for searching point sets in space efficiently.
- */
 template<typename PayloadT = Empty>
 struct Kd_Tree {
     typedef s32 Kd_Node_Idx;
-#  define if_uses_payload if constexpr (!std::is_same_v<PayloadT, Empty>)
-    /**
-     * A node in the k-d-tree. It stores in which axis the space is partitioned (x,y,z)
-     * as an index, the position of the node, and its left and right children.
-     */
+
     struct Kd_Node {
         int axis  =  0;
         V3  point = {};
@@ -63,11 +39,6 @@ struct Kd_Tree {
     Kd_Node*  nodes    = nullptr;
     PayloadT* payloads = nullptr;
 
-    /**
-     * Builds a k-d-tree from the passed point and data array.
-     * @param points The point array.
-     * @param dataArray The data array.
-     */
     static auto build_from(u32 point_count, V3* points, PayloadT* payloads = nullptr) -> Kd_Tree<PayloadT> {
         if (point_count == 0) {
             return {};
@@ -75,8 +46,8 @@ struct Kd_Tree {
 
         Kd_Tree<PayloadT> tree;
 
-        tree.nodes = (Kd_Node*)malloc(sizeof(Kd_Node)  * point_count);
-        if_uses_payload {
+        tree.nodes = (Kd_Node*)malloc(sizeof(Kd_Node) * point_count);
+        if (payloads) {
             tree.payloads = (PayloadT*)malloc(sizeof(payloads) * point_count);
         } else {
             payloads = nullptr;
@@ -84,35 +55,26 @@ struct Kd_Tree {
 
         tree.node_count = 0;
         tree.allocated_node_count = point_count;
-        tree.root = tree._build(points, payloads, 0, 0, point_count);
+        tree.root = tree.build_rec(points, payloads, 0, 0, point_count);
 
         return tree;
     }
 
-
-    /**
-     * Reserves memory for use with @see add.
-     * @param maxNumNodes The maximum number of nodes that can be added using @see addPoint.
-     */
-    void reserve(size_t maxNumNodes) {
+    auto reserve(size_t maxNumNodes) -> void {
         if (allocated_node_count - node_count < maxNumNodes) {
             // at least allocate 16
             allocated_node_count = 2*allocated_node_count < 16 ? 16 : 2*allocated_node_count;
             nodes = (Kd_Node*)realloc(nodes, allocated_node_count * sizeof(Kd_Node));
-            if_uses_payload {
+            if (payloads) {
                 payloads = (PayloadT*)realloc(payloads, allocated_node_count * sizeof(PayloadT));
             }
         }
     }
 
     /**
-     * Adds the passed point and data to the k-d-tree.
-     * WARNING: This function may be less efficient than @see build if the points are added in an order suboptimal
-     * for the search structure. Furthermore, @see reserveDynamic must be called before calling this function.
-     * @param point The point to add.
-     * @param data The corresponding data to add.
+     * Leads to not optimal partitioning
      */
-    void add(V3 point, PayloadT payload = {}) {
+    auto add(V3 point, PayloadT payload = {}) -> void {
         reserve(1);
 
         const int k = 3; // Number of dimensions
@@ -145,57 +107,49 @@ struct Kd_Tree {
         node->right_idx = -1;
 
 
-        if_uses_payload {
+        if (payloads) {
             payloads[node_count] = payload;
         }
 
         node_count++;
     }
 
-    /**
-     * Performs an area search in the k-d-tree and returns all points within a certain bounding box.
-     * @param box The bounding box.
-     * @return The points stored in the k-d-tree inside of the bounding box.
-     */
-    void query_in_aabb(AxisAlignedBox box, Array_List<V3>* out_points, Array_List<PayloadT>* out_payloads = nullptr) {
-        _query_in_aabb(box, out_points, out_payloads, root);
+    auto query_in_aabb(Axis_Aligned_Box box, Array_List<V3>* out_points,
+                       Array_List<PayloadT>* out_payloads = nullptr)
+        -> void
+    {
+        query_in_aabb_rec(box, out_points, out_payloads, root);
     }
 
-
-    /**
-     * Performs an area search in the k-d-tree and returns all points within a certain distance to some center point.
-     * @param centerPoint The center point.
-     * @param radius The search radius.
-     * @return The points stored in the k-d-tree inside of the search radius.
-     */
-    void query_in_sphere(V3 center, f32 radius, Array_List<V3>* out_points, Array_List<PayloadT>* out_payloads = nullptr) {
+    auto query_in_sphere(V3 center, f32 radius, Array_List<V3>* out_points,
+                         Array_List<PayloadT>* out_payloads = nullptr)
+        -> void
+    {
         // First, find all points within the bounding box containing the search circle.
-        AxisAlignedBox box;
+        Axis_Aligned_Box box;
         box.min = center - V3{ radius, radius, radius };
         box.max = center + V3{ radius, radius, radius };
 
         u32 out_points_original_count = out_points->count;
 
-        _query_in_aabb(box, out_points, out_payloads, root);
+        query_in_aabb_rec(box, out_points, out_payloads, root);
 
         // Now, filter all points out that are not within the search radius.
-        float squaredRadius = radius*radius;
-        V3 differenceVector;
+        float squared_radius = radius*radius;
+        V3 diff_vec;
 
         u32 point_idx = out_points_original_count;
         while (point_idx < out_points->count) {
             V3 point = (*out_points)[point_idx];
-            differenceVector = point - center;
+            diff_vec = point - center;
 
-            if (differenceVector.x * differenceVector.x +
-                differenceVector.y * differenceVector.y +
-                differenceVector.z * differenceVector.z > squaredRadius)
+            if (diff_vec.x * diff_vec.x +
+                diff_vec.y * diff_vec.y +
+                diff_vec.z * diff_vec.z > squared_radius)
             {
                 out_points->remove_index(point_idx);
-                if_uses_payload {
-                    if (out_payloads) {
-                        out_payloads->remove_index(point_idx);
-                    }
+                if (payloads && out_payloads) {
+                    out_payloads->remove_index(point_idx);
                 }
                 continue;
             }
@@ -204,14 +158,7 @@ struct Kd_Tree {
         }
     }
 
-    /**
-     * Performs an area search in the k-d-tree and returns the number of points within a certain distance to some
-     * center point.
-     * @param centerPoint The center point.
-     * @param radius The search radius.
-     * @return The number of points stored in the k-d-tree inside of the search radius.
-     */
-    size_t get_count_in_sphere(V3 center, f32 radius) {
+    auto get_count_in_sphere(V3 center, f32 radius) -> u32 {
         Array_List<V3> points_in_sphere;
         points_in_sphere.init();
         defer { points_in_sphere.deinit(); };
@@ -221,28 +168,21 @@ struct Kd_Tree {
         return points_in_sphere.count;
     }
 
-    /**
-     * Returns the nearest neighbor in the k-d-tree to the passed point position.
-     * @param point The point to which to find the closest neighbor to.
-     * @return The closest neighbor within the maximum distance.
-     */
-    void findNearestNeighbor(V3 point, f32* out_neighbor_dist, V3* out_neighbor_pos, PayloadT* out_neighbor_payload = nullptr) {
+    auto find_nearest_neighbor(V3 point, f32* out_neighbor_dist, V3* out_neighbor_pos,
+                               PayloadT* out_neighbor_payload = nullptr)
+        -> void
+    {
         *out_neighbor_dist = FLT_MAX;
-        _findNearestNeighbor(point, out_neighbor_dist, out_neighbor_pos, out_neighbor_payload, root);
+        find_nearest_neighbor_rec(point, out_neighbor_dist, out_neighbor_pos, out_neighbor_payload, root);
     }
 
-
-    /**
-     * Builds a k-d-tree from the passed point array recursively (for internal use only).
-     * @param points The point array.
-     * @param dataArray The data array.
-     * @param depth The current depth in the tree (starting at 0).
-     * @return The parent node of the current sub-tree.
-     */
-    Kd_Node_Idx _build(V3* input_points, PayloadT* input_payloads, int depth, size_t startIdx, size_t endIdx) {
+    auto build_rec(V3* input_points, PayloadT* input_payloads, int depth,
+                   size_t start_idx, size_t edx_idx)
+        -> Kd_Node_Idx
+    {
         const int k = 3; // Number of dimensions
 
-        if (endIdx - startIdx == 0) {
+        if (edx_idx - start_idx == 0) {
             return -1;
         }
 
@@ -250,7 +190,7 @@ struct Kd_Tree {
 
         Kd_Node* node     = &nodes[node_idx];
         PayloadT* payload;
-        if_uses_payload {
+        if (payloads) {
             payload = &payloads[node_idx];
         }
 
@@ -258,17 +198,17 @@ struct Kd_Tree {
 
         int axis = depth % k;
 
-        Array_Description payload_array_desc[] { input_payloads + startIdx, sizeof(input_payloads[0]) };
+        Array_Description payload_array_desc[] { input_payloads + start_idx, sizeof(input_payloads[0]) };
         int array_descriptors;
-        if_uses_payload {
+        if (payloads) {
             array_descriptors = 1;
         } else {
             array_descriptors = 0;
         }
 
         soa_sort_r(
-            {input_points  + startIdx, sizeof(input_points[0])},
-            payload_array_desc, array_descriptors, endIdx-startIdx,
+            {input_points  + start_idx, sizeof(input_points[0])},
+            payload_array_desc, array_descriptors, edx_idx-start_idx,
             [] (const void *v1, const void *v2, void* arg) -> s32 {
                 s32 axis = *(s32*)arg;
                 V3* a = (V3*)v1;
@@ -276,27 +216,24 @@ struct Kd_Tree {
                 return (s32)((*a)[axis] > (*b)[axis]);
             }, &axis);
 
-        size_t medianIndex = startIdx + (endIdx - startIdx) / 2;
+        size_t median_idx = start_idx + (edx_idx - start_idx) / 2;
 
-        if_uses_payload {
-            *payload = input_payloads[medianIndex];
+        if (payloads) {
+            *payload = input_payloads[median_idx];
         }
-        node->axis = axis;
-        node->point = input_points[medianIndex];
-        node->left_idx = _build(input_points, input_payloads, depth + 1, startIdx, medianIndex);
-        node->right_idx = _build(input_points, input_payloads, depth + 1, medianIndex + 1, endIdx);
+
+        node->axis      = axis;
+        node->point     = input_points[median_idx];
+        node->left_idx  = build_rec(input_points, input_payloads, depth + 1, start_idx, median_idx);
+        node->right_idx = build_rec(input_points, input_payloads, depth + 1, median_idx + 1, edx_idx);
         return node_idx;
     }
 
 
-    /**
-     * Performs an area search in the k-d-tree and returns all points within a certain bounding box
-     * (for internal use only).
-     * @param box The bounding box.
-     * @param node The current k-d-tree node that is searched.
-     * @param points The points of the k-d-tree inside of the bounding box.
-     */
-    void _query_in_aabb(AxisAlignedBox& box, Array_List<V3>* out_points,  Array_List<PayloadT>* out_payloads, Kd_Node_Idx node_idx) {
+    auto query_in_aabb_rec(Axis_Aligned_Box& box, Array_List<V3>* out_points,
+                           Array_List<PayloadT>* out_payloads, Kd_Node_Idx node_idx)
+        -> void
+    {
         if (node_idx == -1) {
             return;
         }
@@ -304,29 +241,23 @@ struct Kd_Tree {
         Kd_Node* node = &nodes[node_idx];
         if (box.contains(node->point)) {
             out_points->append(node->point);
-            if_uses_payload {
+            if (payloads) {
                 out_payloads->append(payloads[node_idx]);
             }
         }
 
 
         if (box.min[node->axis] <= node->point[node->axis]) {
-            _query_in_aabb(box, out_points, out_payloads, node->left_idx);
+            query_in_aabb_rec(box, out_points, out_payloads, node->left_idx);
         }
         if (box.max[node->axis] >= node->point[node->axis]) {
-            _query_in_aabb(box, out_points, out_payloads, node->right_idx);
+            query_in_aabb_rec(box, out_points, out_payloads, node->right_idx);
         }
     }
 
-    /**
-     * Returns the nearest neighbor in the k-d-tree to the passed point position.
-     * @param point The point to which to find the closest neighbor to.
-     * @param nearestNeighborDistance The distance to the nearest neighbor found so far.
-     * @param nearestNeighbor The nearest neighbor found so far.
-     * @param node The current k-d-tree node that is searched.
-     */
-    void _findNearestNeighbor(V3 point, f32* out_dist_to_neighbor, V3* out_neighbor_pos,
-                              PayloadT* out_neighbor_payload, Kd_Node_Idx node_idx)
+    auto find_nearest_neighbor_rec(V3 point, f32* out_dist_to_neighbor, V3* out_neighbor_pos,
+                                   PayloadT* out_neighbor_payload, Kd_Node_Idx node_idx)
+        -> void
     {
         if (node_idx == -1) {
             return;
@@ -335,20 +266,20 @@ struct Kd_Tree {
         Kd_Node* node = &nodes[node_idx];
 
         // Descend on side of split planes where the point lies.
-        bool isPointOnLeftSide = point[ node->axis] <= node->point[node->axis];
-        if (isPointOnLeftSide) {
-            _findNearestNeighbor(point, out_dist_to_neighbor, out_neighbor_pos, node->left_idx);
+        bool point_is_on_left = point[ node->axis] <= node->point[node->axis];
+        if (point_is_on_left) {
+            find_nearest_neighbor_rec(point, out_dist_to_neighbor, out_neighbor_pos, node->left_idx);
         } else {
-            _findNearestNeighbor(point, out_dist_to_neighbor, out_neighbor_pos, node->right_idx);
+            find_nearest_neighbor_rec(point, out_dist_to_neighbor, out_neighbor_pos, node->right_idx);
         }
 
         // Compute the distance of this node to the point.
         V3 diff = point - node->point;
-        float newDistance = length(diff);
-        if (newDistance < *out_dist_to_neighbor) {
-            *out_dist_to_neighbor = newDistance;
+        float new_dist = length(diff);
+        if (new_dist < *out_dist_to_neighbor) {
+            *out_dist_to_neighbor = new_dist;
             *out_neighbor_pos     = node->point;
-            if_uses_payload {
+            if (payloads) {
                 if (out_neighbor_payload) {
                     *out_neighbor_payload = payloads[node_idx];
                 }
@@ -356,11 +287,11 @@ struct Kd_Tree {
         }
 
         // Check whether there could be a closer point on the opposite side.
-        if (isPointOnLeftSide && point[node->axis] + *out_dist_to_neighbor >= node->point[node->axis]) {
-            _findNearestNeighbor(point, out_dist_to_neighbor, out_neighbor_pos, node->right_idx);
+        if (point_is_on_left && point[node->axis] + *out_dist_to_neighbor >= node->point[node->axis]) {
+            find_nearest_neighbor_rec(point, out_dist_to_neighbor, out_neighbor_pos, node->right_idx);
         }
-        if (!isPointOnLeftSide && point[node->axis] - *out_dist_to_neighbor <= node->point[node->axis]) {
-            _findNearestNeighbor(point, out_dist_to_neighbor, out_neighbor_pos, node->left_idx);
+        if (!point_is_on_left && point[node->axis] - *out_dist_to_neighbor <= node->point[node->axis]) {
+            find_nearest_neighbor_rec(point, out_dist_to_neighbor, out_neighbor_pos, node->left_idx);
         }
     }
 };
